+++ src/client/js/index.js
@@ -1,3 +1,5 @@
+const io = require('./../../../node_modules/socket.io/client-dist/socket.io.min.js');
+
 import 'bootstrap';
 import '../assets/apple-touch-icon.png';
 import '../assets/favicon.ico';
@@ -14,97 +16,110 @@ const chat = require('./chat'),
 // Initalize user client information
 user.init();
 
-// Create connection to server
-const socket = new WebSocket(
-  '' +
-    (location.protocol == 'https:' ? 'wss' : 'ws') +
-    '://' +
-    location.hostname +
-    (location.port ? ':' + location.port : '') +
-    '?id=' +
-    user.getId()
-);
-
-let interval;
-
-// Heartbeat
-socket.onopen = () => {
-  interval = setInterval(() => {
-    socket.send(JSON.stringify({ type: 'ping', data: true }));
-  }, 10000)
-}
+const socket = io(window.location.origin, {
+  query: {
+    'user-id': user.getId(),
+  }
+});
 
 // Server responses
-socket.onclose = () => {
-  if (interval) clearInterval(interval);
-  user.setPartner(false);
-  toast.toast("You have been disconnected from the server, Please refresh the page.", 'bg-danger');
-  document.getElementById('userCount').innerText = "0";
-};
 
-socket.onerror = (event) => {
-  console.log('Error: ' + event);
-};
+socket.on('connection_success', (id) => {
+  user.setId(id);
+});
 
-socket.onmessage = (event) => {
-  const response = JSON.parse(event.data);
+socket.on('new_session', () => {
+  chat.addChatMessage('You have disconnected from your partner because you open YiffSpot from another place.', {
+    class: 'message-system'
+  });
+  toast.toast('You have been disconnected', 'bg-danger');
+  document.getElementById('userCount').innerText = '0';
+});
 
-  switch(response.type) {
-    case 'connection_success':
-      user.setId(response.data);
-      break;
+socket.on('update_user_count', (count) => {
+  document.getElementById('userCount').innerText = count.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+});
 
-    case 'connection_exists':
-      toast.toast('You already have an active session.');
-      return false;
+socket.on('receive_message', (message) => {
+  chat.addChatMessage(message.content, { class: 'message-partner' + (message.isContributor ? " contributor" : "") });
+});
 
-    case 'update_user_count':
-      document.getElementById('userCount').innerText = response.data.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-      break;
+socket.on('partner_typing', (status) => {
+  if (status == true) chat.showChatTyping();
+  else chat.hideChatTyping();
+});
 
-    case 'receive_message':
-      chat.addChatMessage(response.data, {class: 'message-partner'});
-      break;
+socket.on('partner_connected', (preference) => {
+  partner.connected(preference);
+});
 
-    case 'partner_typing':
-      if (response.data == true) chat.showChatTyping();
-      else chat.hideChatTyping();
-      break;
+socket.on('partner_disconnected', () => {
+  partner.disconnected();
+});
 
-    case 'partner_connected':
-      partner.connected(response.data);
-      break;
+socket.on('partner_left', () => {
+  partner.left();
+});
 
-    case 'partner_disconnected':
-      partner.disconnected();
-      break;
+socket.on('partner_blocked', () => {
+  partner.blocked();
+});
 
-    case 'partner_left':
-      partner.left();
-      break;
+socket.on('partner_pending', () => {
+  partner.pending();
+});
 
-    case 'partner_blocked':
-      partner.blocked();
-      break;
+socket.on('invalid_preferences', () => {
+  preferences.invalid();
+});
 
-    case 'partner_pending':
-      partner.pending();
-      break;
+socket.on('already_finding_partner', () => {
+  chat.addChatMessage('You are already looking for a partner.', {
+    class: 'message-system'
+  });
+});
 
-    case 'invalid_preferences':
-      preferences.invalid();
-      break;
-    case 'client_disconnect':
-      if (!document.getElementById("disconnect-row").classList.contains("hide-ele")) {
-          document.getElementById("disconnect-row").classList.add("hide-ele");
-      }
-      chat.addChatMessage('You have disconnected from your partner.', {
-          class: 'message-system'
-      });
-      user.setPartner(false);
-      break;
+socket.on('preference_updated', () => {
+  chat.addChatMessage('Your preference has been updated.', {
+    class: 'message-system'
+  });
+});
+
+socket.on('client_disconnect', () => {
+  if (!document.getElementById("disconnect-row").classList.contains("hide-ele")) {
+    document.getElementById("disconnect-row").classList.add("hide-ele");
   }
-};
+  chat.addChatMessage('You have disconnected from your partner.', {
+    class: 'message-system'
+  });
+  user.setPartner(false);
+});
+
+socket.on('reconnect_to_partner', (preference) => {
+  chat.showChatBox();
+  partner.reconnect(preference);
+});
+
+socket.on('disconnect', () => {
+  toast.toast('You have been disconnected from the server, Trying to reconnect.', 'bg-danger');
+  document.getElementById('userCount').innerText = "0";
+});
+
+socket.on('contributor_key_accepted', () => {
+  toast.toast('You have successfully tagged as a contributor.');
+});
+
+socket.on('contributor_tag_removed', () => {
+  toast.toast('You have successfully removed your contributor tag.');
+});
+
+window.submitContributorKey = (key) => {
+  socket.emit('send_contributor_key', key);
+}
+
+window.removeContributorTag = () => {
+  socket.emit('remove_contributor_tag');
+}
 
 /**
  * Event Listeners
@@ -124,7 +139,7 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 
   // When key is pressed in message box / typing
-  document.getElementById('messageBox').addEventListener('input', function(e) {
+  document.getElementById('messageBox').addEventListener('input', function (e) {
     e.preventDefault();
     chat.sendTypingStatus(socket);
   });
@@ -136,7 +151,7 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 
   // When block partner button is clicked
-  document.getElementById('block-partner').addEventListener('click', function(e) {
+  document.getElementById('block-partner').addEventListener('click', function (e) {
     e.preventDefault();
     partner.block(socket);
   });
@@ -147,19 +162,43 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 
   // Show Site Settings
-  document.getElementById('settings').addEventListener('click', function(e) {
+  document.getElementById('settings').addEventListener('click', function (e) {
     e.preventDefault();
     document.getElementById('userPrefs').classList.toggle('hide-ele');
     document.getElementById('siteSettings').classList.toggle('hide-ele');
   });
 
   // Show Preferences Settings
-  document.getElementById('preferences').addEventListener('click', function(e) {
+  document.getElementById('preferences').addEventListener('click', function (e) {
     e.preventDefault();
     document.getElementById('siteSettings').classList.toggle('hide-ele');
     document.getElementById('userPrefs').classList.toggle('hide-ele');
   });
 
-  // Handle when window is closed
-  window.addEventListener('beforeunload', () => socket.close());
+  window.onbeforeunload = function () {
+    user.setPartner(false);
+  }
 });
+
+// Gonna hide easter egg cuz yes :)
+console.log(`MMMMMMMMMXklccld0WMMMMMMWXOdooxKWMMMMMMM
+MMMMMMMMNxc:::::lkNMMMMNOl:::::o0WMMMMMM
+MMMMMMMMKo:::::::lOWMMNkc::::::cxNMMMMMM
+MMMMMMMW0l::::::::xXWWOl:::::::cxNMMMMMM
+MMMMMMMMXd::::::::dXWNxc:::::::lOWMMMMMM
+WXK0KXWMW0l::::::ckNWNxc::::::cxNMMMMMMM
+Ooc:clx0NW0oc:::cdXWMW0o::::clkNMMWNNNWM
+l::::::ckNWN0kxk0NMMMMWKkdodkKWWX0xoooxK
+c:::::::ckNMMMMMMMMMMMMMWWWWMMW0oc:::::o
+o::::::::oKMMMMMWWNNNWWMMMMMMWOl:::::::c
+0o:::::::oKMMWN0kdooodxOXWMMMXo::::::::l
+WKxl:::clONWXkoc:::::::clxKNW0o:::::::ck
+MMWXOkkOKNXklcodddlccldddolxKXxc::::coON
+MMMMMMMMN0occdO000OkkO000OdcoOKOdodxOXWM
+MMMMMMMNkc::cxO00000000000xc:lkXWNWWMMMM
+MMMMMMXxc::::lxkO000000Okdl:::ckNMMMMMMM
+MMMMMNkc::::::cloxOOOOxolc:::::l0WMMMMMM
+MMMMMKo::::::::::cldolc::::::::ckWMMMMMM
+MMMMMXxc:::::::ccloooolcc:::::cdKWMMMMMM
+MMMMMWXkolcccloxk0KXNNKOxoc:cokXWMMMMMMM
+=============== YIFF SPOT ==============`);