+++ src/client/js/utility.js
@@ -1,0 +1,9 @@
+const tagsToReplace = {
+    '&': '&amp;',
+    '<': '&lt;',
+    '>': '&gt;'
+};
+
+const replaceTag = (tag) => tagsToReplace[tag] || tag;
+
+exports.safe_tags_replace = (str) => str.replace(/[&<>]/g, replaceTag);