+++ src/server/disconnect.js
@@ -12,12 +12,14 @@ module.exports = (users, token) => {
     // Disconnect user from partner.
     users.removePartner(currentUser.id);
 
-    if (partner.socket.readyState == 1) {
-      partner.socket.send(JSON.stringify({type: 'partner_disconnected', data: true}));
+    if (partner.socket.isAlive) {
+      partner.socket.emit('partner_disconnected');
     }
   }
 
   // Remove disconnected user from clients list
   users.removeClient(currentUser.id);
   users.decrementOnline();
+
+  console.log('User Disconnected... Total Users Online: %d', users.getOnline());
 }